+++ src/ai/backend/manager/data/image/BUILD
@@ -1,0 +1,1 @@
+python_sources(name="src")
