+++ src/ai/backend/manager/services/image/actions/dealias_image.py
@@ -1,9 +1,10 @@
 from dataclasses import dataclass
 from typing import Optional, override
+from uuid import UUID
 
 from ai.backend.manager.actions.action import BaseActionResult
 from ai.backend.manager.actions.exceptions import BaseActionException
-from ai.backend.manager.models.image import ImageAliasRow
+from ai.backend.manager.data.image.types import ImageAliasData
 from ai.backend.manager.services.image.base import ImageAction
 
 
@@ -22,11 +23,12 @@ class DealiasImageAction(ImageAction):
 
 @dataclass
 class DealiasImageActionResult(BaseActionResult):
-    image_alias: ImageAliasRow
+    image_id: UUID
+    image_alias: ImageAliasData
 
     @override
     def entity_id(self) -> Optional[str]:
-        return str(self.image_alias.image_id)
+        return str(self.image_id)
 
 
 class DealiasImageActionNoSuchAliasError(BaseActionException):
