+++ src/ai/backend/manager/services/image/actions/rescan_images.py
@@ -1,9 +1,8 @@
 from dataclasses import dataclass
 from typing import Optional, override
 
-from ai.backend.common.types import DispatchResult
 from ai.backend.manager.actions.action import BaseActionResult
-from ai.backend.manager.models.image import ImageRow
+from ai.backend.manager.data.image.types import RescanImagesResult
 from ai.backend.manager.services.image.base import ImageAction
 
 
@@ -24,8 +23,8 @@ class RescanImagesAction(ImageAction):
 # TODO: BatchAction으로 업데이트, entity_ids는 image row ids로.
 @dataclass
 class RescanImagesActionResult(BaseActionResult):
-    # TODO: DispatchResult 제거하고 list[ImageRow]는 별도의 dataclass로 변경
-    result: DispatchResult[list[ImageRow]]
+    # TODO: result란 이름 사용 지양
+    result: RescanImagesResult
 
     @override
     def entity_id(self) -> Optional[str]:
